+++ R/class_attributes.R
@@ -43,19 +43,8 @@ get_values <- function(object, ...) {
   return(value_data)
 }
 
+# old name: unexported
 get_value <- get_values
-
-get_id_var <- function(object, ...) {
-  id_var <- attributes(object)$id_var
-  return(id_var)
-}
-
-get_id <- function(object, ...) {
-  id_var_name <- object |> get_id_var()
-  id_data <- object |> pull(id_var_name)
-  return(id_data)
-}
-
 
 #' Extract biomarker levels
 #'
@@ -77,33 +66,6 @@ get_biomarker_levels <- function(object, ...) {
     cli::cli_abort("biomarkers not found in `object`")
   }
 }
-
-#' Get biomarker variable name
-#'
-#' @param object a `pop_data` object
-#' @param ... unused
-#'
-#' @returns
-#' a [character] string identifying the biomarker names column in `object`
-#' @export
-#'
-#' @examples
-#' sees_pop_data_100 |> get_biomarker_names_var()
-get_biomarker_names_var <- function(object, ...) {
-  # get value attribute
-  biomarker_var <- attributes(object)[["biomarker_var"]]
-
-  return(biomarker_var)
-}
-
-get_biomarker_names <- function(object, ...) {
-  # get biomarker name data
-  biomarker_names_var <- get_biomarker_names_var(object)
-  biomarker_data <- object |> pull(biomarker_names_var)
-
-  return(biomarker_data)
-}
-
 
 set_age <- function(object,
                     age = "Age",
@@ -194,51 +156,6 @@ set_value <- function(object,
 
     # set id attribute
     attr(object, "value_var") <- "value"
-  }
-
-  return(object)
-}
-
-set_id <- function(object,
-                   id = "index_id",
-                   standardize = TRUE,
-                   ...) {
-  # check if id column exists
-  if (id %in% colnames(object)) {
-    attr(object, "id_var") <- id
-  } else {
-    cli::cli_warn("The specified {.var id} column {.val {id}} does not exist.")
-
-    # search id variable from object
-    id_var <-
-      grep(
-        x = colnames(object),
-        value = TRUE,
-        pattern = id,
-        ignore.case = TRUE
-      )
-
-    if (length(id_var) == 1) {
-      attr(object, "id_var") <- id_var
-
-      # create warning when using searched id instead of provided id
-      cli::cli_inform('Proceeding to use "{id_var}"')
-    } else if (length(id_var) == 0) {
-      cli::cli_abort("No similar column name was detected.")
-    } else {
-      # if (length(id_var) > 1)
-      cli::cli_warn("Multiple potential matches found: {.var {id_var}}")
-      cli::cli_inform("Using first match: {.var {id_var[1]}}")
-      attr(object, "id_var") <- id_var[1]
-    }
-  }
-
-  if (standardize) {
-    object <- object |>
-      rename(c("id" = attr(object, "id_var")))
-
-    # set id attribute
-    attr(object, "id_var") <- "id"
   }
 
   return(object)