+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 * Rename `est.incidence()` to `estimate_scr()` (#389)
+* Improved warning messages for `get_biomarker_names_var()`
 * Added `get_*()` extractor functions to API (#380)
 * Added optional CI error bars to `autoplot.summary.seroincidence.by()` (#372)
 * Improved y-limit calculation in `graph.curve.params()` (#368)
@@ -164,13 +165,13 @@ to avoid printing an OK message.
 
   - `set_age()`
   - `set_value()`
-  - `set_id()`
+  - `set_id_var()`
   - `get_age()`
   - `get_values()`
-  - `get_id()`
+  - `ids()`
   - `get_age_var()`
   - `get_values_var()`
-  - `get_id_var()`
+  - `ids_varname()`
   
 * Added additional warnings to `load_pop_data()`